Refactored the Revision components and redirect to the home page at the end of a revision.
