+++ app/src/pages/RevisionPage.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 
 import { useParams } from "react-router-dom";
-import RevisionContent from "../components/Revision/RevisionContent";
+import RevisionManager from "../components/Revision/RevisionManager";
 import SideMenuLayout from "../components/Layouts/SideMenuLayout";
 
 export default function RevisionPage() {
   const { collectionId } = useParams();
   return (
     <SideMenuLayout>
-      <RevisionContent collectionId={collectionId} />
+      <RevisionManager collectionId={collectionId} />
     </SideMenuLayout>
   );
 }
